+++ Farmer/utils/Keybindings.lua
@@ -1,7 +1,7 @@
 local addonName, addon = ...;
 
 function addon.exposeHeader (headerName, text)
-  local bindingToken = "BINDING_HEADER_FARMER_" .. headerName;
+  local bindingToken = 'BINDING_HEADER_FARMER_' .. headerName;
 
   assert(_G[bindingToken] == nil,
     addonName .. ': binding already exists: "' .. headerName .. '"');
@@ -10,13 +10,14 @@ function addon.exposeHeader (headerName, text)
 end
 
 function addon.exposeBinding(bindingName, text, handler)
-  local bindingToken = "BINDING_NAME_FARMER_" .. bindingName;
+  local bindingToken = 'BINDING_NAME_FARMER_' .. bindingName;
+  local handlerToken = 'FARMER_' .. bindingName;
 
   assert(_G[bindingToken] == nil,
-    addonName .. ': binding already exists: "' .. bindingName .. '"');
+      addonName .. ': binding already exists: "' .. bindingName .. '"');
+  assert(_G[handlerToken] == nil,
+      addonName .. ': handler already exists: "' .. bindingName .. '"');
 
   _G[bindingToken] = text;
-  _G[bindingToken .. '_HANDLER'] = handler;
+  _G[handlerToken] = handler;
 end
-
-addon.exposeHeader('HEADER', addonName);